Move PatternSet factory implementation
