+++ subprojects/core-api/src/main/java/org/gradle/api/tasks/util/internal/DefaultPatternSetFactory.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2025 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.api.tasks.util.internal;
+
+import org.gradle.api.tasks.util.PatternSet;
+import org.gradle.internal.Factory;
+
+public class DefaultPatternSetFactory implements Factory<PatternSet> {
+
+    private final PatternSpecFactory patternSpecFactory;
+
+    public DefaultPatternSetFactory(PatternSpecFactory patternSpecFactory) {
+        this.patternSpecFactory = patternSpecFactory;
+    }
+
+    @Override
+    public PatternSet create() {
+        return new InternalPatternSet(patternSpecFactory);
+    }
+
+    // This is only required to avoid adding a new public constructor to the public `PatternSet` type.
+    private static class InternalPatternSet extends PatternSet {
+        public InternalPatternSet(PatternSpecFactory patternSpecFactory) {
+            super(patternSpecFactory);
+        }
+    }
+}
